+++ base/cvd/cuttlefish/host/commands/assemble_cvd/BUILD.bazel
@@ -332,6 +332,7 @@ cf_cc_library(
         "//cuttlefish/host/graphics_detector:graphics_detector_cc_proto",
         "//cuttlefish/host/libs/config:config_constants",
         "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/host/libs/config:guest_hwui_renderer",
         "//cuttlefish/host/libs/config:vmm_mode",
         "//libbase",
         "@fmt",
