+++ base/cvd/cuttlefish/host/commands/assemble_cvd/graphics_flags.cc
@@ -30,6 +30,7 @@
 #include "cuttlefish/host/graphics_detector/graphics_detector.pb.h"
 #include "cuttlefish/host/libs/config/config_constants.h"
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
+#include "cuttlefish/host/libs/config/guest_hwui_renderer.h"
 #include "cuttlefish/host/libs/config/vmm_mode.h"
 
 #ifdef __APPLE__
